+++ urbanairship-react-native/ios/UARCTModule/UARCTModuleVersion.m
@@ -4,7 +4,7 @@
 
 @implementation UARCTModuleVersion
 
-NSString *const airshipModuleVersionString = @"14.4.4";
+NSString *const airshipModuleVersionString = @"14.5.0";
 
 + (nonnull NSString *)get {
     return airshipModuleVersionString;